==> R/2_Whale_risk.Rmd <==
---
title: "Whale risk - Mn3kmSDM"
author: "Sam Woodman"
date: '`r format(Sys.Date(), "%d %B %Y")`'
output:
  pdf_document: default
  html_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

NOTE: This script is based on the file of the same name in https://github.com/jameals/raimbow/tree/master/humpback_risk

## Introduction

The code in this document calculates (humpback) whale risk of entanglement for each grid cell as: humpback density * fishing measure (e.g. # of VMS pings or $ of Dungeness crab landed). Thus, the units of the calculated risk values are (humpback density * VMS measure) / km^2, with the following notes:

1) the VMS measure is VMS pings.

2) this document uses non-confidential (i.e., only cells with >= 3 pings) VMS data for Oregon and Washington and confidential (i.e. all) data for California.

```{r, message=FALSE}
library(tidyverse)
library(lubridate)
library(sf)
library(here)
```

Set file paths depending on user.

```{r}
flag.save <- TRUE

file.data.mnpreds <- here("data", "products", "Humpback_5km_long_monthly.rds")
file.data.grid <- here("data", "5x5 km grid shapefile", "five_km_grid_polys_geo.shp")
file.landerased <- here("data", "Grid_5km_landerased.rds")

file.vms <- file.path(
  "C:/SMW/SMW-Misc/Forney-dynamic-mn/raimbow-local", 
  "Data/Non-confidential VMS monthly data/", 
  "Non-confidential VMS data summarized by 5km grid cell 2009-18 All 3 states.csv")
file.vms.ca <- file.path(
  "C:/SMW/SMW-Misc/Forney-dynamic-mn/raimbow-local", 
  "Data/fishDataCA/CA_DCRB_vms_fishing_daily_2009-2019_all_vessels.RDS")

# file.out.risk <- "C:/SMW/RAIMBOW/raimbow-local/RDATA_files/Whale_risk.Rdata"
file.risk.long <- here("data", "products", "Whale_risk_long_nona.rds")
```

## User options

The user should update this section as desired to specify 1) the dates for which risk is calculated and 2) the specify the fishing measure, e.g. the number of VMS pings.

```{r}
# 1) Dates for which to calculate risk
# df.key.ym <- tibble(
#   year = c(rep(2015, 2), sapply(2016:2017, rep, 12)), 
#   month = sprintf("%02d", c(11, 12, rep(1:12, 2)))
# )
# # Below is example df.key.ym for Nov 2009 - Dec 2018
# df.key.ym <- tibble(
#   year = c(rep(2009, 2), sapply(2010:2018, rep, 12)),
#   month = sprintf("%02d", c(11, 12, rep(1:12, 9)))
# )

df.key.ym <- tibble(
  year = c(rep(2009, 2), sapply(2010:2018, rep, 12), rep(2019, 7)),
  month = sprintf("%02d", c(11, 12, rep(1:12, 9), 1:7))
)

### 2) Specify how to represent fishing effort, ie the column name from file.vms
# vms.colname <- "dollars_DCRB_noncon"
vms.colname <- "Num_DCRB_VMS_pings"
vms.colname.all <- "Num_DCRB_VMS_pings_noncon"
```

## Read and process humpback predictions

First we read in the humpback predictions that have been aggregated (averaged) by month. We will use the density values to calculate risk, so we do not calculate the abundance values. 

```{r}
grid.5km.geom <- st_read(file.data.grid)

humpback.dens.long <- readRDS(file.data.mnpreds)
humpback.dens <- humpback.dens.long %>%
  mutate(year_mo = paste("H_Avg", year(date), sprintf("%02d", month(date)), sep = "_")) %>% 
  select(GRID5KM_ID, area_km_lno, year_mo, Humpback_dens_mean) %>% 
  spread(key = year_mo, value = Humpback_dens_mean) %>%
  left_join(st_set_geometry(grid.5km.geom, NULL), by = "GRID5KM_ID") %>%
  select(GRID5KM_ID, LONGITUDE, LATITUDE, area_km_lno, starts_with("H_avg"))
```

## Read and process fishing data

```{r}
###############################################################################
# Fishing data processing

#------------------------------------------------------------------------------
### Process fishing data - CA
fish.daily.ca <- readRDS(file.vms.ca)
fish.summ.ca <- fish.daily.ca %>% 
  rename(VMS_curr = !!vms.colname) %>%
  mutate(VMS_metric = suppressWarnings(as.numeric(VMS_curr)),
         year_mo = paste0("DC_", year_month)) %>%
  group_by(year_mo, GRID5KM_ID) %>%
  summarise(year = unique(year),
            VMS_sum = sum(VMS_metric, na.rm = TRUE), 
            .groups = "drop") %>%
  left_join(grid.5km.geom, by = "GRID5KM_ID") %>% 
  filter(LATITUDE < 42) %>% 
  select(year, year_mo, GRID5KM_ID, VMS_sum)

### Read and process other (all) data - OR and WA
fish.summ.all <- read_csv(file.vms, show_col_types = FALSE) %>%
  rename(VMS_curr = !!vms.colname.all) %>%
  mutate(VMS_metric = suppressWarnings(as.numeric(VMS_curr)),
         year_mo = paste0("DC_", year_mo)) %>%
  group_by(year_mo, GRID5KM_ID) %>%
  summarise(year = unique(year),
            VMS_sum = sum(VMS_metric, na.rm = TRUE), 
            .groups = "drop") %>%
  left_join(grid.5km.geom, by = "GRID5KM_ID") %>% 
  filter(LATITUDE >= 42,
         !(GRID5KM_ID %in% fish.summ.ca$GRID5KM_ID)) %>% 
  select(year, year_mo, GRID5KM_ID, VMS_sum)

### Join CA and other data
fish.summ <- bind_rows(fish.summ.ca, fish.summ.all) %>%
  arrange(year_mo) %>% 
  select(year_mo, GRID5KM_ID, VMS_sum) %>%
  pivot_wider(id_cols = GRID5KM_ID, names_from = year_mo, values_from = VMS_sum) %>% 
  arrange(GRID5KM_ID)

### Add grid info
fish.out <- humpback.dens %>% 
  select(GRID5KM_ID, LONGITUDE, LATITUDE) %>% 
  left_join(fish.summ, by = "GRID5KM_ID") %>%
  select(GRID5KM_ID, LONGITUDE, LATITUDE, starts_with("DC"))
# fish.out <- fish.summ %>% 
#   left_join(st_set_geometry(grid.5km.geom, NULL), by = "GRID5KM_ID") %>%
#   select(GRID5KM_ID, LONGITUDE, LATITUDE, starts_with("DC"))

stopifnot(!any(duplicated(fish.out$GRID5KM_ID)))
# d <- fish.out %>% filter(!(GRID5KM_ID %in% humpback.dens$GRID5KM_ID))
# sum(select(d, starts_with("DC_")), na.rm = TRUE)
```

## Calculate risk

We create the data frame to use as a key for getting data form the desired months. Then we loop through the key, calculating and storing the risk as we go.

```{r}
### Add Mn and fish data column names to data frame
df.key <- df.key.ym %>% 
  mutate(humpback_name = paste("H_Avg", year, month, sep = "_"), 
         fish_name = paste("DC", year, month, sep = "_"))

risk.all <- humpback.all <- fish.all <- 
  humpback.dens %>% select(GRID5KM_ID, LONGITUDE, LATITUDE, area_km_lno)
name.idx <- ncol(risk.all)

#------------------------------------------------------------------------------
### For loop through year-month pairs
for (i in 1: nrow(df.key)) {
  # Conditional for months where there is no fishing data, and thus no column
  #   in fishing data frame
  if (df.key$fish_name[i] %in% names(fish.out)) {
    df.out <- data.frame(
      humpback.dens %>% select(GRID5KM_ID, df.key$humpback_name[i]), 
      fish.out %>% select(df.key$fish_name[i])
    )
  } else {
    df.out <- data.frame(
      humpback.dens %>% select(GRID5KM_ID, df.key$humpback_name[i]), 
      as.numeric(NA)
    )
  }
  
  # Calculate risk, and add data to applicable data frames
  df.out <- df.out %>% 
    set_names(c("GRID5KM_ID", "humpback_curr", "fish_curr")) %>% 
    mutate(risk_curr = humpback_curr * fish_curr)
  
  risk.all <- cbind(risk.all, df.out$risk_curr)
  names(risk.all)[i+name.idx] <- paste("Mn_DC_risk", df.key$year[i], df.key$month[i], sep = "_")
  
  humpback.all <- cbind(humpback.all, df.out$humpback_curr)
  names(humpback.all)[i+name.idx] <- paste("Mn", df.key$year[i], df.key$month[i], sep = "_")
  
  fish.all <- cbind(fish.all, df.out$fish_curr)
  names(fish.all)[i+name.idx] <- paste("DC", df.key$year[i], df.key$month[i], sep = "_")
  
  
  # file.out <- paste0("Mn_DC_risk", df.key$year[i], "_", df.key$month[i], ".csv")
  # write.csv(df.out, file = file.out, row.names = FALSE)
  
  rm(df.out) #, file.out)
}; rm(i, name.idx)
```

Now we can save the key, humpback densities, fishing values, and risk values for use in other scripts. 

```{r}
# Save files for Whale_risk_maps
risk.all.dens <- risk.all
fish.all.num <- fish.all
humpback.all.dens <- humpback.all

# Convert humpback and risk values to abundance and total, respectively
mult.func <- function(x, y) {x * y}
grid.area <- humpback.all.dens %>% select(GRID5KM_ID, area_km_lno)

humpback.abund <- humpback.all.dens %>% 
  mutate(across(.cols = starts_with("Mn_"), 
                \(x) mult.func(x, y = grid.area$area_km_lno)))
fish.pings <- fish.all.num
risk.total <- risk.all.dens %>% 
  mutate(across(.cols = starts_with("Mn_DC_risk_"), 
                \(x) mult.func(x, y = grid.area$area_km_lno)))

# if (flag.save)
#   save(df.key, humpback.abund, fish.pings, risk.total, file = file.out.risk)
```

## Prep data frames for time series analyses

For time series analyses, we need long data frames that of the humpback abundance, VMS pings, and total risk (whales * VMS pings). We also remove the rows with all NA risk values - time series plots much have these removed.

```{r}
### Monthly
risk.nona.any <- apply(select(risk.total, starts_with("Mn_DC_risk")), 1, function(i) any(!is.na(i))) 
fish.nona.any <- apply(select(fish.all.num, starts_with("DC_")), 1, function(i) any(!is.na(i)))

# Fishing data has smaller spatial footprint, and thus fish NAs should match risk NAs
#   However, it shouldn't be a show-stopper if this isn't true
identical(risk.nona.any, fish.nona.any)

# Will do all plotting with nona data frames for ease of computation, 
#   but mostly for summing data only across cells that have risk
risk.nona <- risk.total[risk.nona.any, ]
humpback.nona <- humpback.abund[risk.nona.any, ]
fish.nona <- fish.all.num[risk.nona.any, ]
```

## Data processing

The first step is to make the whale, VMS, and risk data frames 'long', meaning there is one row for each grid cell for each time period

```{r}
h.df <- humpback.nona %>% 
  gather(col_name, mn_abund, 
         -GRID5KM_ID, -LONGITUDE, -LATITUDE, -area_km_lno) %>% 
  mutate(yr = as.numeric(substr(col_name, 4, 7)), 
         mon = as.numeric(substr(col_name, 9, 10)), 
         ym = paste(yr, sprintf("%02d", mon), sep = "-")) %>% 
  select(GRID5KM_ID, LONGITUDE, LATITUDE, area_km_lno, ym, yr, mon, mn_abund)

f.df <- fish.nona %>% 
  gather(col_name, vms_pings, 
         -GRID5KM_ID, -LONGITUDE, -LATITUDE, -area_km_lno) %>% 
  mutate(yr = as.numeric(substr(col_name, 4, 7)), 
         mon = as.numeric(substr(col_name, 9, 10))) %>% 
  select(-col_name)

r.df <- risk.nona %>% 
  gather(col_name, risk_total, 
         -GRID5KM_ID, -LONGITUDE, -LATITUDE, -area_km_lno) %>% 
  mutate(yr = as.numeric(substr(col_name, 12, 15)), 
         mon = as.numeric(substr(col_name, 17, 18))) %>% 
  select(-col_name)

# Sanity check
col.same <- c("GRID5KM_ID", "LONGITUDE", "LATITUDE", "area_km_lno", "yr", "mon")
stopifnot(
  all.equal(h.df[, col.same], f.df[, col.same]), 
  all.equal(h.df[, col.same], r.df[, col.same])
)

# 
all.df <- h.df %>% 
  left_join(f.df, by = col.same) %>% 
  left_join(r.df, by = col.same)


if (flag.save) saveRDS(all.df, file = file.risk.long)
```
